refactor: collapse synthetic modules into hl7v2
