+++ crates/hl7v2-corpus/Cargo.toml
@@ -13,14 +13,11 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-sha2 = "0.10.9"
-chrono = { version = "0.4.44", features = ["serde", "std", "clock"], default-features = false }
-thiserror = { workspace = true }
-rand = "0.10.0"
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "synthetic",
+] }
 
 [dev-dependencies]
+hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
 proptest = "1.6"
-
+serde_json = { workspace = true }
