+++ crates/hl7v2-gen/Cargo.toml
@@ -13,19 +13,9 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-# Use microcrates directly for finer-grained dependencies
-hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
-hl7v2-writer = { version = "1.2.0", path = "../hl7v2-writer" }
-# Keep hl7v2-core for backward compatibility
-hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
-# ACK generation is now in its own crate
-hl7v2-ack = { version = "1.2.0", path = "../hl7v2-ack" }
-# Faker functionality is now in its own crate
-hl7v2-faker = { version = "1.2.0", path = "../hl7v2-faker" }
-# Template functionality is now in its own crate
-hl7v2-template = { version = "1.2.0", path = "../hl7v2-template" }
-serde = { workspace = true, features = ["derive"] }
-serde_yaml = { package = "serde_yaml_ng", version = "0.10.0" }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "synthetic",
+] }
 
 [dev-dependencies]
 proptest = "1.6"
@@ -33,6 +23,7 @@ rand = "0.10.0"
 cucumber = "0.22.1"
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 toml = "0.8"
+hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
 hl7v2-query = { version = "1.2.0", path = "../hl7v2-query" }
 hl7v2-normalize = { version = "1.2.0", path = "../hl7v2-normalize" }
 hl7v2-parser = { version = "1.2.0", path = "../hl7v2-parser" }
