+++ crates/hl7v2-gen/src/lib.rs
@@ -1,46 +1,5 @@
-//! Deterministic HL7 v2 message generator.
+//! Deprecated compatibility crate.
 //!
-//! This crate provides functionality for generating synthetic HL7 v2
-//! messages based on templates and profiles.
-//!
-//! # Template-Based Generation
-//!
-//! Template-based message generation functionality is available through
-//! the [`hl7v2_template`] crate and re-exported here for convenience.
-//! See the [`hl7v2_template`] documentation for details on template
-//! structure and value sources.
-//!
-//! # ACK Generation
-//!
-//! ACK (acknowledgment) generation functionality is available through
-//! the [`hl7v2_ack`] crate and re-exported here for convenience.
-//!
-//! # Faker Data Generation
-//!
-//! Realistic test data generation (names, addresses, medical codes, etc.)
-//! is available through the [`hl7v2_faker`] crate and re-exported here
-//! for convenience.
-//!
-//! # Example
-//!
-//! ```
-//! use hl7v2_gen::{Template, generate, ack, AckCode, Faker, FakerValue};
-//! ```
+//! Use `hl7v2::synthetic::generate` instead.
 
-// Re-export template functionality from hl7v2-template crate for backward compatibility
-pub use hl7v2_template::{
-    Template, ValueSource, generate, generate_corpus, generate_distributed_corpus,
-    generate_diverse_corpus, generate_golden_hashes, verify_golden_hashes,
-};
-
-// Re-export ACK functionality from hl7v2-ack crate for backward compatibility
-pub use hl7v2_ack::{AckCode, ack, ack_with_error};
-
-// Re-export faker functionality from hl7v2-faker crate for backward compatibility
-pub use hl7v2_faker::{DateError, Faker, FakerValue, GaussianError, GenerateError};
-
-// Re-export core types that are commonly used with this crate
-pub use hl7v2_core::{Atom, Comp, Delims, Error, Field, Message, Rep, Segment};
-
-#[cfg(test)]
-mod tests;
+pub use hl7v2::synthetic::generate::*;
