+++ crates/hl7v2-template-values/Cargo.toml
@@ -13,19 +13,17 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
-hl7v2-faker = { version = "1.2.0", path = "../hl7v2-faker" }
-chrono = { version = "0.4.44", features = ["serde", "std", "clock"], default-features = false }
-rand = "0.10.0"
-serde = { workspace = true, features = ["derive"] }
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "synthetic",
+] }
 
 [dev-dependencies]
 cucumber = "0.22.1"
 proptest = "1.6.0"
+rand = "0.10.0"
 serde_json = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 
 [[test]]
 name = "bdd_tests"
 harness = false
-
