+++ crates/hl7v2-template/Cargo.toml
@@ -13,16 +13,10 @@ keywords.workspace = true
 categories.workspace = true
 
 [dependencies]
-hl7v2-model = { version = "1.2.0", path = "../hl7v2-model" }
-hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
-hl7v2-template-values = { version = "1.2.0", path = "../hl7v2-template-values" }
-hl7v2-corpus = { version = "1.2.0", path = "../hl7v2-corpus" }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-rand = "0.10.0"
-sha2 = "0.10.9"
-
+hl7v2 = { version = "1.2.0", path = "../hl7v2", default-features = false, features = [
+    "synthetic",
+] }
 
 [dev-dependencies]
+hl7v2-core = { version = "1.2.0", path = "../hl7v2-core" }
 proptest = "1.6"
-
