+++ crates/hl7v2/src/synthetic.rs
@@ -1,0 +1,7 @@
+//! Synthetic message generation APIs.
+
+pub mod corpus;
+pub mod faker;
+pub mod generate;
+pub mod template;
+pub mod values;
